+++ backend/src/main/java/org/eclipse/tractusx/puris/backend/common/edc/domain/repository/SingleLevelBomAsPlannedContractMappingRepository.java
@@ -1,6 +1,5 @@
 /*
- * Copyright (c) 2025 Volkswagen AG
- * Copyright (c) 2025 Contributors to the Eclipse Foundation
+ * Copyright (c) 2026 Volkswagen AG
  *
  * See the NOTICE file(s) distributed with this work for additional
  * information regarding copyright ownership.
@@ -17,15 +16,15 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  */
-package org.eclipse.tractusx.puris.backend.masterdata.domain.repository;
+package org.eclipse.tractusx.puris.backend.common.edc.domain.repository;
+import org.eclipse.tractusx.puris.backend.common.edc.domain.model.SingleLevelBomAsPlannedContractMapping;
+import org.springframework.stereotype.Repository;
 
-import java.util.List;
-import java.util.UUID;
+@Repository
+public interface SingleLevelBomAsPlannedContractMappingRepository extends GeneralContractMappingRepository<SingleLevelBomAsPlannedContractMapping> {
 
-import org.eclipse.tractusx.puris.backend.masterdata.domain.model.MaterialRelation;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface MaterialRelationRepository extends JpaRepository<MaterialRelation, UUID> {
-
-	List<MaterialRelation> findAllByParentOwnMaterialNumber(String parentOwnMaterialNumber);
+    @Override
+    default Class<? extends SingleLevelBomAsPlannedContractMapping> getType() {
+        return SingleLevelBomAsPlannedContractMapping.class;
+    }
 }